feat(page.tsx): add metadata for code of conduct page to improve SEO and page information The metadata is now included for the code of conduct page using the `getPageMetadata` function, which enhances SEO and provides better page information.
fix(sitemap.ts): correct root path in sitemap to ensure proper URL structure
The root path in the sitemap is corrected from an empty string to '/', ensuring the correct URL structure for the sitemap.

chore(code-of-conduct): update submodule reference to latest commit
The submodule for the code of conduct is updated to the latest commit, ensuring the latest changes are included.