chore(gitignore): add .wrangler to ignore list
Add .wrangler to the .gitignore file to prevent unnecessary files
from being tracked in the repository. This helps maintain a clean
working directory by excluding configuration files specific to
Wrangler, a tool for managing Cloudflare Workers.

docs(code-of-conduct): update code of conduct with minor wording
adjustment and update last updated date

Enhance the clarity of the code of conduct by adding "display" to
the list of prohibited actions regarding illegal activities. This
ensures the document is comprehensive in its guidelines. Update the
last updated date to reflect the latest changes, ensuring users are
aware of the most current version.